flight: Move test auth handlers to test_auth_handlers.h,cc
